Improvised SUMAS Excel result sheet template
This temporary improvised Excel result template is for SUMAS lecturers . In the view of the Programmer, it is best suited for those lecturers that handle small classes. That is class where the maximum number of students are less than the maximum number in the result list. However, the main goal is to assist those that handle general classes to meet up with the timeline often given in our Institution for result submission.

NB: Using and applying this result template is the users choice. Observe that it satisfied your expectation before using it. The programmer/author will no be liable to anything fringe of whatever right or standard existing and can be attributed to product of technology or innovation as can be applicable to this result sheet template.

However, the programmer is open for constructive suggestion, progressive contribution to this work as it will be best for the generality of the lecturers to individual perculiarities.

For further inquiry on how to use this Template, please consult any Computer science lecturer or any ICT staff or any staff who has basic Computer Appreciation skill.

Enjoy SUMAS improvised technology for solving our local problem.